Remove logic from view functions
